Remove cold storage cleanup while preserving joined database shutdown
